server/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
Markdown==3.7
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
uuid==1.30
whitenoise==6.7.0
